+++ LICENSE
@@ -2,6 +2,8 @@ MIT License
 
 Copyright (c) 2024 Quozul
 
+(Copyright (c) 2026 Pelijot)
+
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
 in the Software without restriction, including without limitation the rights
