+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'PterodactylPowerAction'
+rootProject.name = 'CraftyPowerAction'
