--- src/main/java/fr/pickaria/pterodactylpoweraction/OnlineChecker.java
@@ -1,9 +1,0 @@
-package fr.pickaria.pterodactylpoweraction;
-
-import java.util.concurrent.CompletableFuture;
-
-public interface OnlineChecker {
-    CompletableFuture<Void> waitForRunning();
-
-    boolean isRunningNow();
-}
