--- src/main/java/fr/pickaria/pterodactylpoweraction/api/WhitelistEntry.java
@@ -1,3 +1,0 @@
-package fr.pickaria.pterodactylpoweraction.api;
-
-public record WhitelistEntry(String uuid, String name) {}
