--- src/main/java/fr/pickaria/pterodactylpoweraction/configuration/APIType.java
@@ -1,6 +1,0 @@
-package fr.pickaria.pterodactylpoweraction.configuration;
-
-public enum APIType {
-    PTERODACTYL,
-    SHELL,
-}
