--- src/main/java/fr/pickaria/pterodactylpoweraction/configuration/PingMethod.java
@@ -1,6 +1,0 @@
-package fr.pickaria.pterodactylpoweraction.configuration;
-
-public enum PingMethod {
-    PING,
-    PTERODACTYL,
-}
