--- src/main/java/fr/pickaria/pterodactylpoweraction/configuration/ShutdownBehaviour.java
@@ -1,7 +1,0 @@
-package fr.pickaria.pterodactylpoweraction.configuration;
-
-public enum ShutdownBehaviour {
-    SHUTDOWN_EMPTY,
-    SHUTDOWN_ALL,
-    NOTHING,
-}
