+++ src/main/java/observer/pelijot/craftypoweraction/OnlineChecker.java
@@ -1,0 +1,9 @@
+package observer.pelijot.craftypoweraction;
+
+import java.util.concurrent.CompletableFuture;
+
+public interface OnlineChecker {
+    CompletableFuture<Void> waitForRunning();
+
+    boolean isRunningNow();
+}
