+++ src/main/java/observer/pelijot/craftypoweraction/PowerActionAPI.java
@@ -1,0 +1,11 @@
+package observer.pelijot.craftypoweraction;
+
+import java.util.concurrent.CompletableFuture;
+
+public interface PowerActionAPI {
+    CompletableFuture<Void> stop(String server);
+
+    CompletableFuture<Void> start(String server);
+
+    CompletableFuture<Boolean> isPlayerWhitelisted(String server, String playerName);
+}
