+++ src/main/java/observer/pelijot/craftypoweraction/api/WhitelistEntry.java
@@ -1,0 +1,3 @@
+package observer.pelijot.craftypoweraction.api;
+
+public record WhitelistEntry(String uuid, String name) {}
