+++ src/main/java/observer/pelijot/craftypoweraction/configuration/APIType.java
@@ -1,0 +1,6 @@
+package observer.pelijot.craftypoweraction.configuration;
+
+public enum APIType {
+    CRAFTY,
+    SHELL,
+}
