+++ src/main/java/observer/pelijot/craftypoweraction/configuration/PingMethod.java
@@ -1,0 +1,6 @@
+package observer.pelijot.craftypoweraction.configuration;
+
+public enum PingMethod {
+    PING,
+    CRAFTY,
+}
