+++ src/main/java/observer/pelijot/craftypoweraction/configuration/ShutdownBehaviour.java
@@ -1,0 +1,7 @@
+package observer.pelijot.craftypoweraction.configuration;
+
+public enum ShutdownBehaviour {
+    SHUTDOWN_EMPTY,
+    SHUTDOWN_ALL,
+    NOTHING,
+}
