--- src/main/templates/fr/pickaria/pterodactylpoweraction/BuildConstants.java
@@ -1,7 +1,0 @@
-package fr.pickaria.pterodactylpoweraction;
-
-// The constants are replaced before compilation
-public class BuildConstants {
-
-    public static final String VERSION = "${version}";
-}
