+++ src/main/templates/observer/pelijot/craftypoweraction/BuildConstants.java
@@ -1,0 +1,7 @@
+package observer.pelijot.craftypoweraction;
+
+// The constants are replaced before compilation
+public class BuildConstants {
+
+    public static final String VERSION = "${version}";
+}
